sym link to git config

+++ install.ps1
@@ -26,14 +26,17 @@ npm install -g @devcontainers/cli
 Copy-Item -Force -Path "$PSScriptRoot\wsl\.wslconfig" -Destination "$env:USERPROFILE\.wslconfig"
 
 # Symbolic link to git config
-Remove-Item -Path "$env:USERPROFILE\.gitconfig" -ErrorAction Ignore
-New-Item -Path "$env:USERPROFILE\.gitconfig" -ItemType SymbolicLink -Target "$env:USERPROFILE\OneDrive - Microsoft\.gitconfig"
+if (Test-Path -Path "$env:ONEDRIVE\.gitconfig") {
+    Write-Output "Creating symbolic link to git config file in OneDrive..."
+    Remove-Item -Path "$env:USERPROFILE\.gitconfig" -ErrorAction Ignore
+    New-Item -Path "$env:USERPROFILE\.gitconfig" -ItemType SymbolicLink -Target "$env:ONEDRIVE\.gitconfig"
+} else {
+    Write-Warning "OneDrive git config file not found. Skipping symbolic link creation."
+}
 
 . "$PSScriptRoot\PowerShell\Remove-DesktopShortcuts.ps1"
 
-# Run git\configure_git.sh
 # Create symlink to gh config file in WSL instance
-# Write-Verbose -Message "Creating symlink in WSL instance $wslInstanceName to the gh cli config file..."
 # Remove the existing gh config file if it exists.
 # Invoke-Wsl $wslInstanceName "[ -d .config/gh ] && rm -r .config/gh" | Write-Verbose
 # Create a symlink to the Windows gh config file.
